Add building instructions.
Add the current instructions for building the plug-in.  These have been verified to work, but are not ideal (old Visual Studio required).  Newer Visual Studio will likely build the plug-in, but will need additional support (e.g. .NET framework) installed on user's systems.  This needs to be handled somehow in the future.